Run a batch of reports.

Run one report multiple times with different sets of parameters.

orderly_batch(name = NULL, parameters = NULL, ...)

Arguments

name
Name of the report to run (see orderly_list()). A leading src/ will be removed if provided, allowing easier use of autocomplete. Alternatively, the default of NULL is useful if you have already set the working directory to be the source directory.

parameters
Data frame of parameters passed to report. Each row represents a parameter set to be passed to one report run.

...
Additional args passed to orderly_run()

Value

List of ids of newly created reports

See also

Examples

path <- orderly::orderly_example("demo")
params <- data.frame(nmin = c(0.2, 0.25))
ids <- orderly::orderly_batch("other", params, root = path)

#> [ name       ]  other

#> [ id         ]  20210714-153224-dc6f2cf4

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0.2

#> [ start      ]  2021-07-14 15:32:24

#> [ data       ]  source => extract: 19 x 2

#> [ parameter  ]  nmin: 0.2

#> 
#> > extract$number <- extract$number + rnorm(1)
#> 
#> > write.csv(extract, "summary.csv", row.names = TRUE)
#> 
#> > png("graph.png")

#> Warning: unable to load shared object '/usr/local/lib/R/modules//R_X11.so':
#>   libXt.so.6: cannot open shared object file: No such file or directory

#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > do_plot(extract)

#> 
#> > dev.off()
#> agg_png 
#>       2 

#> [ end        ]  2021-07-14 15:32:24

#> [ elapsed    ]  Ran report in 0.06315851 secs

#> [ artefact   ]  summary.csv: 08a4566d063098080bfd318f675926f2
#> [ ...        ]  graph.png: 9d98e56ce7726a1523c1b03eb0e4efb4

#> [ name       ]  other

#> [ id         ]  20210714-153224-fa3c13d1

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0.25

#> [ start      ]  2021-07-14 15:32:24

#> [ data       ]  source => extract: 19 x 2

#> [ parameter  ]  nmin: 0.25

#> 
#> > extract$number <- extract$number + rnorm(1)
#> 
#> > write.csv(extract, "summary.csv", row.names = TRUE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > do_plot(extract)

#> 
#> > dev.off()
#> agg_png 
#>       2 

#> [ end        ]  2021-07-14 15:32:25

#> [ elapsed    ]  Ran report in 0.01683092 secs

#> [ artefact   ]  summary.csv: 96f402853ce06d5ef44a98c9cdb67573
#> [ ...        ]  graph.png: f6870aa327d5504aa4bfded46f169df6