Pack and import bundles with remotes

Pack a bundle on a remote. This is like calling orderly_bundle_pack() on the remote and can be used to extract a long-running report from a server to run (say) on an HPC system.

orderly_bundle_pack_remote(
  name,
  parameters = NULL,
  instance = NULL,
  root = NULL,
  locate = TRUE,
  remote = NULL,
  dest = tempdir()
)

orderly_bundle_import_remote(path, root = NULL, locate = TRUE, remote = NULL)

Arguments

name
Name of the report to pack (see orderly_list(). A leading src/ will be removed if provided, allowing easier use of autocomplete.

parameters
Parameters passed to the report. A named list of parameters declared in the orderly.yml. Each parameter must be a scalar character, numeric, integer or logical.

instance
Select instance of the source database to be used, where multiple instances are configured. Use a single unnamed character string to indicate an instance to match. If given, then this name must be present in all databases where instances are listed in orderly_config.yml, and will be ignored by all database where instances are not given. See the "orderly" vignette for further information.

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

remote
The remote to pack the bundle from, or import into

dest
Optional path to write bundle to (a directory name). By default we use the temporary directory and return the full path to the created file.

path
The path to unpack and import (a zip file created by orderly_bundle_run)

Details

The workflow here will typically be:

Use orderly_bundle_pack_remote() to create a local copy of a bundle, extracted from a remote. Typically this will be run from the system where the bundle will be run (an HPC head-node or another powerful computer).

Run the bundle using orderly_bundle_run()

Re-import the completed bundle using orderly_bundle_import_remote which sends the zip file to the remote and adds it to the archive.

Typically these commands will not be run from the orderly root. However, the root argument may still be used to find your remote configuration. Alternatively, if your remote argument is an orderly remote (e.g., orderly_remote_path(), or orderlyweb's orderlyweb::orderlyweb_remote) then the root and locate arguments will be ignored and this command can be run from anywhere. This is the recommended configuration for running on a HPC system.