Commit a generated report

Commit a generated report, moving it from the draft/ directory to archive/ and updating the orderly index. Once committed, reports should not be deleted.

orderly_commit(id, name = NULL, root = NULL, locate = TRUE)

Arguments

id
The identifier of the report

name
The name of the report - this can be omitted and the name will be determined from the id.

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

Value

The path to the newly committed report

Examples

# In a new example orderly, run a report
path <- orderly::orderly_example("minimal")
id <- orderly::orderly_run("example", root = path)

#> [ name       ]  example

#> [ id         ]  20200302-113139-ad001704

#> [ data       ]  source => dat: 20 x 2

#> [ start      ]  2020-03-02 11:31:39

#> 
#> > png("mygraph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(dat$number, dat$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2020-03-02 11:31:39

#> [ elapsed    ]  Ran report in 0.01184487 secs

#> [ artefact   ]  mygraph.png: 5f8e1c71677703144642256f70c86502

# To commit it, all we need is the report id
orderly::orderly_commit(id, root = path)

#> [ commit     ]  example/20200302-113139-ad001704

#> [ copy       ]

#> [ import     ]  example:20200302-113139-ad001704

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3d60f1fae6/archive/example/20200302-113139-ad001704"

# The report is now committed, and as such could be used as a
# depenency in another report and is not subject to deletion by
# orderly::orderly_cleanup
orderly::orderly_list_archive(root = path)

#>      name                       id
#> 1 example 20200302-113139-ad001704