Retrieve orderly config object.

Retrieve orderly config object.

orderly_config(root = NULL, locate = TRUE)

Arguments

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

Value

An R6 object representing the orderly config.

Public fields

root

Root dir of the orderly repository

raw

The raw orderly config yaml

destination

DB connection configuration for where to store orderly output database. Defaults to local SQLite db orderly.sqlite

fields

Configuration of fields in reports, specifying which are required

remote

Configuration of remote sources i.e. shared copy of orderly on a remote machine

vault

Vault server connection information

global_resources

Path to dir containing global resources.

changelog

Changelog type configuration

tags

List of available tags for orderly reports.

database

Database configuration specifying driver and connection args for (possibly multiple) databases

archive_version

Orderly version number of the archive

run_options

List of run options

Methods

Public methods

orderly_config_$new()

orderly_config_$server_options()

orderly_config_$add_run_option()

orderly_config_$get_run_option()

Method new()

Create an object representing orderly config

Usage

orderly_config_$new(root, validate = TRUE)

Arguments

root

Root dir of the orderly repository

validate

If TRUE migrate cfg to handle any format changes and validate structure if well formed for each of the cfg fields

Method server_options()

Get connection options for the current server. This is the details from the "remote" section for the server being run on. Server identified via env var ORDERLY_API_SERVER_IDENTITY

Usage

orderly_config_$server_options()

Returns

Options for current server if can be identified, otherwise NULL

Method add_run_option()

Add a key-value pair run option

Usage

orderly_config_$add_run_option(name, value)

Arguments

name

Name of run option

value

Value for run option

Method get_run_option()

Retrieve value of a run option

Usage

orderly_config_$get_run_option(name)

Arguments

name

Name of run option

Examples

# The orderly demo, with lots of potential reports:
path <- orderly::orderly_example("demo")

orderly::orderly_config(path)

#> <orderly_config>
#>   Public:
#>     add_run_option: function (name, value) 
#>     archive_version: 1.1.25
#>     changelog: data.frame
#>     database: list
#>     destination: list
#>     fields: data.frame
#>     get_run_option: function (name) 
#>     global_resources: global
#>     initialize: function (root, validate = TRUE) 
#>     raw: list
#>     remote: NULL
#>     root: /tmp/RtmpYrLUBu/file4659fc3d45
#>     run_options: list
#>     server_options: function () 
#>     tags: dataset plot
#>     vault: NULL
#>   Private:
#>     migrate: function () 
#>     validate: function ()