Return info about a report which has been run

This will return info from either successful or failed reports. It will look for the report with id in archive first and then look in drafts if it can't be found from archive.

orderly_info(id, name, root = NULL, locate = TRUE)

Arguments

id
The report ID

name
The name of the report

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

Value

Info from report run - this is subject to change. Returns a list which includes report id, name, indication of success, run date and elapsed time, parameters, git info (if available), path to logfile (if exists) and details of error if the run failed

Examples

path <- orderly::orderly_example("demo")
id <- orderly::orderly_run("minimal", root = path)

#> [ name       ]  minimal

#> [ id         ]  20210714-153229-f254e75c

#> [ start      ]  2021-07-14 15:32:29

#> [ data       ]  source => dat: 20 x 2

#> 
#> > png("mygraph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(dat$number, dat$name), las = 2)

#> 
#> > dev.off()
#> agg_png 
#>       2 

#> [ end        ]  2021-07-14 15:32:29

#> [ elapsed    ]  Ran report in 0.01555443 secs

#> [ artefact   ]  mygraph.png: 175369b2bcf4115f343c8ad746c0c072

orderly::orderly_info(id, "minimal", root = path)

#> $id
#> [1] "20210714-153229-f254e75c"
#> 
#> $name
#> [1] "minimal"
#> 
#> $success
#> [1] TRUE
#> 
#> $date
#> [1] "2021-07-14 15:32:29"
#> 
#> $elapsed
#> [1] 0.01555443
#> 
#> $git
#> NULL
#> 
#> $parameters
#> NULL
#> 
#> $logfile
#> NULL
#> 
#> $error
#> NULL
#>