Find most recent report

Find most recent version of an orderly report. The most recent report is always the most recently run report that has been committed (regardless of the order in which they were committed).

orderly_latest(
  name = NULL,
  root = NULL,
  locate = TRUE,
  draft = FALSE,
  must_work = TRUE
)

Arguments

name
Name of the report to find; if NULL returns the most recent report across all names

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

draft
Should draft reports be used searched? Valid values are logical (TRUE, FALSE) or use the string newer to use draft reports where they are newer than archive reports. For consistency, always and never are equivalent to TRUE and FALSE, respectively.

must_work
Throw an error if no report is found. If FALSE, returns NA_character_.

Value

A character string with the id of the most recent report

See also

Examples

path <- orderly::orderly_example("minimal")
id1 <- orderly::orderly_run("example", root = path, echo = FALSE)

#> [ name       ]  example

#> [ id         ]  20210714-153230-a823442b

#> [ start      ]  2021-07-14 15:32:30

#> [ data       ]  source => dat: 20 x 2

#> [ end        ]  2021-07-14 15:32:30

#> [ elapsed    ]  Ran report in 0.01574612 secs

#> [ artefact   ]  mygraph.png: 175369b2bcf4115f343c8ad746c0c072

id2 <- orderly::orderly_run("example", root = path, echo = FALSE)

#> [ name       ]  example

#> [ id         ]  20210714-153230-b7ce57aa

#> [ start      ]  2021-07-14 15:32:30

#> [ data       ]  source => dat: 20 x 2

#> [ end        ]  2021-07-14 15:32:30

#> [ elapsed    ]  Ran report in 0.01473498 secs

#> [ artefact   ]  mygraph.png: 175369b2bcf4115f343c8ad746c0c072

# With no reports committed there is no latest report:
orderly::orderly_latest("example", root = path, must_work = FALSE)

#> [1] NA

# Commit the first report and it will be reported as latest:
orderly::orderly_commit(id1, root = path)

#> [ commit     ]  example/20210714-153230-a823442b

#> [ copy       ]

#> [ import     ]  example:20210714-153230-a823442b

#> [ success    ]  :)

#> [1] "/tmp/RtmpYrLUBu/file4616db5bf1/archive/example/20210714-153230-a823442b"

orderly::orderly_latest("example", root = path)

#> [1] "20210714-153230-a823442b"

# Commit the second report and it will be reported as latest instead:
orderly::orderly_commit(id2, root = path)

#> [ commit     ]  example/20210714-153230-b7ce57aa

#> [ copy       ]

#> [ import     ]  example:20210714-153230-b7ce57aa

#> [ success    ]  :)

#> [1] "/tmp/RtmpYrLUBu/file4616db5bf1/archive/example/20210714-153230-b7ce57aa"

orderly::orderly_latest("example", root = path)

#> [1] "20210714-153230-b7ce57aa"