Migrate orderly2 sources

Migrate source code from orderly2 (version 1.99.81 and previous) to refer to orderly (version 1.99.82 and subsequent). This is a one-off and will not exist in the package for long, get it while it's hot. This function does a very simple-minded replacement of orderly2 with orderly in R files (extension .R or .r), markdown (.md), R markdown (.Rmd) and quarto (.qmd). It requires a clean git status before it is run, and will be best to run against a fresh clone of a repository. After running, review changes (if any) with git diff and then commit.

Usage

orderly_migrate_source_from_orderly2(path = ".", dry_run = FALSE)

Arguments

path

Path to the repo. We will not change anything here unless the path is under source control, and unless the git status is "up to date" (i.e., no local unsaved modifications or untracked files). It is recommended to run this against a fresh clone.

dry_run

Logical, indicating if no changes would be made, but just print information about the changes that would be made.

Value

Primarily called for side effects, but returns (invisibly) TRUE if any changes were made, FALSE otherwise.