Orderly remote at a different path

Create a "handle" for interacting with orderly repositories that are hosted at a different path. This might be useful in cases where you have access to an orderly repository via a network mount or a synchronised folder (e.g., Dropbox, Box, etc). More generally, orderly_remote_path implements an interface used by orderly to abstract over different ways that orderly repositories might be hosted remotely, including over HTTP APIs.

orderly_remote_path(path, name = NULL)

Arguments

path
Path to the orderly store

name
Name of the remote

See also

Examples

# Suppose we have a "remote" orderly repository at some path.
# This might be read-only for you in practice and available via a
# network filesystem or a dropbox folder synced to your computer.
# We'll populate this with a pair of reports:
path_remote <- orderly::orderly_example("demo")
id <- orderly::orderly_run("other", list(nmin = 0),
                           root = path_remote, echo = FALSE)

#> [ name       ]  other

#> [ id         ]  20190813-162457-cfed24ab

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0

#> [ data       ]  source => extract: 20 x 2

#> [ start      ]  2019-08-13 16:24:57

#> [ end        ]  2019-08-13 16:24:57

#> [ elapsed    ]  Ran report in 0.01295328 secs

#> [ artefact   ]  summary.csv: 23877598325ab7d45999349ef868a7f5
#> [ ...        ]  graph.png: 7360cb2eed3327ff8a677b3598ed7343

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  other/20190813-162457-cfed24ab

#> [ copy       ]

#> [ import     ]  other:20190813-162457-cfed24ab

#> [ success    ]  :)

#> [1] "/tmp/RtmpC4lmv4/filed34a64d2a/archive/other/20190813-162457-cfed24ab"

id <- orderly::orderly_run("use_dependency",
                           root = path_remote, echo = FALSE)

#> [ name       ]  use_dependency

#> [ id         ]  20190813-162458-1893b0f7

#> [ depends    ]  other@20190813-162457-cfed24ab:summary.csv -> incoming.csv

#> [ start      ]  2019-08-13 16:24:58

#> [ end        ]  2019-08-13 16:24:58

#> [ elapsed    ]  Ran report in 0.01268435 secs

#> [ artefact   ]  graph.png: 7360cb2eed3327ff8a677b3598ed7343

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  use_dependency/20190813-162458-1893b0f7

#> [ copy       ]

#> [ import     ]  use_dependency:20190813-162458-1893b0f7

#> [ success    ]  :)

#> [1] "/tmp/RtmpC4lmv4/filed34a64d2a/archive/use_dependency/20190813-162458-1893b0f7"

# We'll create a an object to interact with this remote using
# orderly_remote_path.
remote <- orderly::orderly_remote_path(path_remote)

# We can use this object directly
remote$list_reports()

#>  [1] "changelog"          "connection"         "global"            
#>  [4] "html"               "interactive"        "minimal"           
#>  [7] "multi-artefact"     "multifile-artefact" "other"             
#> [10] "spaces"             "use_dependency"     "use_resource"      
#> [13] "use_resource_dir"   "view"              

remote$list_versions("other")

#> [1] "20190813-162457-cfed24ab"

# More typically one will interact with the functions
# orderly_pull_archive and orderly_pull_dependencies.

# Now, suppose that you have your "local" copy of this; it shares
# the same source (ordinarily these would both be under version
# control with git):
path_local <- orderly::orderly_example("demo")

# If we wanted to run the report "use_dependency" we need to have
# a copy of the report "other", on which it depends:
try(orderly::orderly_run("use_dependency", root = path_local))

#> [ name       ]  use_dependency

#> Error in orderly_latest(name, config, FALSE, draft = draft, must_work = must_work) : 
#>   Did not find any archive reports for other

# We can "pull" depenencies of a report before running
orderly::orderly_pull_dependencies("use_dependency", remote = remote,
                                   root = path_local)

#> [ depends    ]  use_dependency has 1 dependency

#> [ pull       ]  other:20190813-162457-cfed24ab

#> [ import     ]  other:20190813-162457-cfed24ab

# Now we can run the report because we have a local copy of the
# dependency:
orderly::orderly_run("use_dependency", root = path_local)

#> [ name       ]  use_dependency

#> [ id         ]  20190813-162458-83b8c49e

#> [ depends    ]  other@20190813-162457-cfed24ab:summary.csv -> incoming.csv

#> [ start      ]  2019-08-13 16:24:58

#> 
#> > d <- read.csv("incoming.csv", stringsAsFactors = FALSE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(d$number, d$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2019-08-13 16:24:58

#> [ elapsed    ]  Ran report in 0.01915026 secs

#> [ artefact   ]  graph.png: 7360cb2eed3327ff8a677b3598ed7343

#> [1] "20190813-162458-83b8c49e"

# We can also directly pull previously run reports:
orderly::orderly_pull_archive("use_dependency", id, remote = remote,
                              root = path_local)

#> [ pull       ]  use_dependency:20190813-162458-1893b0f7

#> [ depends    ]  other/20190813-162457-cfed24ab

#> [ pull       ]  other:20190813-162457-cfed24ab already exists, skipping

#> [ import     ]  use_dependency:20190813-162458-1893b0f7

orderly::orderly_list_archive(root = path_local)

#>             name                       id
#> 1          other 20190813-162457-cfed24ab
#> 2 use_dependency 20190813-162458-1893b0f7