Run a batch of reports.

Run one report multiple times with different sets of parameters.

orderly_batch(name = NULL, parameters = NULL, ...)

Arguments

name
Name of the report to run (see orderly_list). A leading src/ will be removed if provided, allowing easier use of autocomplete. Alternatively, the default of NULL is useful if you have already set the working directory to be the source directory.

parameters
Data frame of parameters passed to report. Each row represents a parameter set to be passed to one report run.

...
Additional args passed to orderly_run

Value

List of ids of newly created reports

See also

Examples

path <- orderly::orderly_example("demo")
params <- data.frame(nmin = c(0.2, 0.25))
ids <- orderly::orderly_batch("other", params, root = path)

#> [ name       ]  other

#> [ id         ]  20200403-103534-e2e1f2aa

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0.2

#> [ data       ]  source => extract: 19 x 2

#> [ start      ]  2020-04-03 10:35:34

#> 
#> > extract$number <- extract$number + rnorm(1)
#> 
#> > write.csv(extract, "summary.csv", row.names = TRUE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > do_plot(extract)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2020-04-03 10:35:35

#> [ elapsed    ]  Ran report in 0.2171006 secs

#> [ artefact   ]  summary.csv: da1bdb89343fc0c1c9b19f3d257d27f7
#> [ ...        ]  graph.png: 0f5857002852d4d85f0d3423907eeb2c

#> [ name       ]  other

#> [ id         ]  20200403-103535-3f324f9d

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0.25

#> [ data       ]  source => extract: 19 x 2

#> [ start      ]  2020-04-03 10:35:35

#> 
#> > extract$number <- extract$number + rnorm(1)
#> 
#> > write.csv(extract, "summary.csv", row.names = TRUE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > do_plot(extract)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2020-04-03 10:35:35

#> [ elapsed    ]  Ran report in 0.02500963 secs

#> [ artefact   ]  summary.csv: 8d5a99a3a15cb4d7225059854b331981
#> [ ...        ]  graph.png: 0add2123b9b0e96c744f83678226bab3