Connect to orderly databases

Connect to the orderly databases. These should be treated as as read-only.

orderly_db(type, root = NULL, locate = TRUE, validate = TRUE, instance = NULL)

Arguments

type
The type of connection to make (source, destination, csv or rds).

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

validate
Logical, indicating if the database schema should be validated on open (currently only applicable with type = "destination"). This is primarily intended for internal use.

instance
Used only by type = "source", and used to select the instance, where multiple instances are configured. Use a single unnamed character string to indicate an instance to match. If given, then this name must be present in all databases where instances are listed in orderly_config.yml, and will be ignored by all database where instances are not given. See the "orderly" vignette for further information.

Value

A database connection, or list of connections in the case of source.

Details

Orderly has several databases:

source

All of the databases named in the database section of the orderly_config.yml

destination

The orderly index database (typically a SQLite database stored at the orderly root)

csv

The cache of database query results, in csv format

rds

The cache of database query results, in rds format

Examples

# Create an orderly that has a single commited report:
path <- orderly::orderly_example("minimal")
id <- orderly::orderly_run("example", root = path)

#> [ name       ]  example

#> [ id         ]  20200302-113139-fe2b5d11

#> [ data       ]  source => dat: 20 x 2

#> [ start      ]  2020-03-02 11:31:40

#> 
#> > png("mygraph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(dat$number, dat$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2020-03-02 11:31:40

#> [ elapsed    ]  Ran report in 0.01157665 secs

#> [ artefact   ]  mygraph.png: 5f8e1c71677703144642256f70c86502

orderly::orderly_commit(id, root = path)

#> [ commit     ]  example/20200302-113139-fe2b5d11

#> [ copy       ]

#> [ import     ]  example:20200302-113139-fe2b5d11

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3d5872dcd1/archive/example/20200302-113139-fe2b5d11"

# The source database holds the data that might be accessible via
# the 'data' entry in orderly.yml:
db <- orderly::orderly_db("source", root = path)
# This is a list, with one connection per database listed in the
# orderly_config.yml (an empty list if none are specified):
db

#> $source
#> <SQLiteConnection>
#>   Path: /tmp/Rtmp3R7mks/file3d5872dcd1/source.sqlite
#>   Extensions: TRUE
#> 

DBI::dbListTables(db$source)

#> [1] "data"  "thing"

head(DBI::dbReadTable(db$source, "data"))

#>   id thing      value
#> 1  1    12 -0.3899086
#> 2  2     6 -0.4908328
#> 3  3     8 -1.0457177
#> 4  4     7 -0.8962113
#> 5  5    11  1.2693872
#> 6  6    17  0.5938409

DBI::dbDisconnect(db$source)

# The destination database holds information about the archived
# reports:
db <- orderly::orderly_db("destination", root = path)
DBI::dbListTables(db)

#>  [1] "artefact_format"              "changelog"                   
#>  [3] "changelog_label"              "custom_fields"               
#>  [5] "data"                         "depends"                     
#>  [7] "file"                         "file_artefact"               
#>  [9] "file_input"                   "file_input_global"           
#> [11] "file_purpose"                 "orderly_schema"              
#> [13] "orderly_schema_tables"        "parameters"                  
#> [15] "parameters_type"              "report"                      
#> [17] "report_version"               "report_version_artefact"     
#> [19] "report_version_custom_fields" "report_version_data"         
#> [21] "report_version_package"       "report_version_view"         

# These tables are documented online:
# https://vimc.github.io/orderly/schema
DBI::dbReadTable(db, "report_version")

#>                         id  report                date displayname description
#> 1 20200302-113139-fe2b5d11 example 2020-03-02 11:31:40        <NA>        <NA>
#>   connection published git_sha git_branch git_clean
#> 1          0         0    <NA>       <NA>        NA