Connect to orderly databases

Connect to the orderly databases. These should be treated as as read only (with the exception of source).

orderly_db(type, root = NULL, locate = TRUE, validate = TRUE)

Arguments

type
The type of connection to make (source, destination, csv or rds).

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE).

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

validate
Logical, indicating if the database schema should be validated on open (currently only applicable with type = "destination"). This is primarily intended for internal use.

Examples

# Create an orderly that has a single commited report:
path <- orderly::orderly_example("minimal")
id <- orderly::orderly_run("example", root = path)

#> [ name       ]  example

#> [ id         ]  20190813-162451-8a444c82

#> [ data       ]  source => dat: 20 x 2

#> [ start      ]  2019-08-13 16:24:51

#> 
#> > png("mygraph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(dat$number, dat$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2019-08-13 16:24:51

#> [ elapsed    ]  Ran report in 0.0123415 secs

#> [ artefact   ]  mygraph.png: 7360cb2eed3327ff8a677b3598ed7343

orderly::orderly_commit(id, root = path)

#> [ commit     ]  example/20190813-162451-8a444c82

#> [ copy       ]

#> [ import     ]  example:20190813-162451-8a444c82

#> [ success    ]  :)

#> [1] "/tmp/RtmpC4lmv4/filed6717f4b3/archive/example/20190813-162451-8a444c82"

# The source database holds the data that might be accessible via
# the 'data' entry in orderly.yml:
db <- orderly::orderly_db("source", root = path)
# This is a list, with one connection per database listed in the
# orderly_config.yml (an empty list if none are specified):
db

#> $source
#> <SQLiteConnection>
#>   Path: /tmp/RtmpC4lmv4/filed6717f4b3/source.sqlite
#>   Extensions: TRUE
#> 

DBI::dbListTables(db$source)

#> [1] "data"  "thing"

head(DBI::dbReadTable(db$source, "data"))

#>   id thing       value
#> 1  1    19  0.06016044
#> 2  2     6 -0.58889449
#> 3  3    10  0.53149619
#> 4  4     7 -1.51839408
#> 5  5    14  0.30655786
#> 6  6     6 -1.53644982

DBI::dbDisconnect(db$source)

# The destination database holds information about the archived
# reports:
db <- orderly::orderly_db("destination", root = path)
DBI::dbListTables(db)

#>  [1] "artefact_format"              "changelog"                   
#>  [3] "changelog_label"              "custom_fields"               
#>  [5] "data"                         "depends"                     
#>  [7] "file"                         "file_artefact"               
#>  [9] "file_input"                   "file_purpose"                
#> [11] "orderly_schema"               "orderly_schema_tables"       
#> [13] "parameters"                   "parameters_type"             
#> [15] "report"                       "report_version"              
#> [17] "report_version_artefact"      "report_version_custom_fields"
#> [19] "report_version_data"          "report_version_package"      
#> [21] "report_version_view"         

# These tables are documented online:
# https://vimc.github.io/orderly/schema
DBI::dbReadTable(db, "report_version")

#>                         id  report                date displayname description
#> 1 20190813-162451-8a444c82 example 2019-08-13 16:24:51        <NA>        <NA>
#>   connection published git_sha git_branch git_clean
#> 1          0         0    <NA>       <NA>        NA