Set default remote location

Set and get default remote locations. Default locations are specific to an orderly repository (based on the path of the repository) so there is no interaction between different orderly projects.

orderly_default_remote_set(value, root = NULL, locate = TRUE)

orderly_default_remote_get(root = NULL, locate = TRUE)

Arguments

value
A string describing a remote, a remote object, or NULL to clear

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

Value

The default remote (for orderly_default_remote_get). The function orderly_default_remote_set is called for its side effects only.

Examples

# Same setup as in orderly_remote_path, with a remote orderly:
path_remote <- orderly::orderly_example("demo")
id <- orderly::orderly_run("other", list(nmin = 0),
                           root = path_remote, echo = FALSE)

#> [ name       ]  other

#> [ id         ]  20200302-113142-2069bf25

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0

#> [ data       ]  source => extract: 20 x 2

#> [ start      ]  2020-03-02 11:31:42

#> [ end        ]  2020-03-02 11:31:42

#> [ elapsed    ]  Ran report in 0.01081276 secs

#> [ artefact   ]  summary.csv: 01c70f6e51fab381154775c7e70ed7b3
#> [ ...        ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  other/20200302-113142-2069bf25

#> [ copy       ]

#> [ import     ]  other:20200302-113142-2069bf25

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3da053dab/archive/other/20200302-113142-2069bf25"

id <- orderly::orderly_run("use_dependency",
                           root = path_remote, echo = FALSE)

#> [ name       ]  use_dependency

#> [ id         ]  20200302-113142-58982e80

#> [ depends    ]  other@20200302-113142-2069bf25:summary.csv -> incoming.csv

#> [ start      ]  2020-03-02 11:31:42

#> [ end        ]  2020-03-02 11:31:42

#> [ elapsed    ]  Ran report in 0.01260567 secs

#> [ artefact   ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3
#> [ ...        ]  info.rds: d463abbd52634836996451f5ae95baf3

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  use_dependency/20200302-113142-58982e80

#> [ copy       ]

#> [ import     ]  use_dependency:20200302-113142-58982e80

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3da053dab/archive/use_dependency/20200302-113142-58982e80"

# And a local orderly
path_local <- orderly::orderly_example("demo")

# We'll create an object to interact with this remote using
# orderly_remote_path.
remote <- orderly::orderly_remote_path(path_remote)

# There is no remote set by default:
try(orderly::orderly_default_remote_get(root = path_local))

#> Error in orderly::orderly_default_remote_get(root = path_local) : 
#>   default remote has not been set yet: use 'orderly::orderly_default_remote_set'

# We can set one:
orderly::orderly_default_remote_set(remote, root = path_local)

# and now we can retrieve it:
orderly::orderly_default_remote_get(root = path_local)

#> <orderly_remote_path>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     config: orderly_config
#>     initialize: function (path, name) 
#>     list_reports: function () 
#>     list_versions: function (name) 
#>     name: /tmp/Rtmp3R7mks/file3da053dab
#>     pull: function (name, id) 
#>     run: function (...) 
#>     url_report: function (name, id) 

# Note that this has not affected the other orderly:
try(orderly::orderly_default_remote_get(root = path_remote))

#> Error in orderly::orderly_default_remote_get(root = path_remote) : 
#>   default remote has not been set yet: use 'orderly::orderly_default_remote_set'