Set up an orderly example

Set up one of the orderly examples included with the package. These are not intended to be starting points for new orderly repositories, but are used in the package examples and vignettes.

orderly_example(name, path = tempfile(), run_demo = FALSE, quiet = FALSE)

Arguments

name
Name of the example

path
Destination to create the example - if it exists already it must be an empty directory. By default, creates a new temporary directory

run_demo
Logical, indicating if the example is configured as a "demo" (i.e., with a set of reports to be run and committed), should these be run?

quiet
Logical, indicating if informational messages should be suppressed when running the demo.

Value

Returns the path to the orderly example

Examples

# Create a new copy of the "minimal" example
path <- orderly::orderly_example("minimal")
dir(path)

#> [1] "archive"            "data"               "demo.yml"          
#> [4] "draft"              "orderly_config.yml" "README.md"         
#> [7] "source.sqlite"      "src"               

# Example reports within this repository:
orderly::orderly_list(path)

#> [1] "example"