Print the dependency tree for a given report using orderly log

orderly allows a report to rely on the artefacts of one or more other orderly reports. This allows users to develop a network of interconnected reports where the output from report becomes the source of data for another. There are two natural questions that can develop around this workflow:

orderly_graph(
  name,
  id = "latest",
  root = NULL,
  locate = TRUE,
  direction = "downstream",
  propagate = TRUE,
  max_depth = 100,
  show_all = FALSE
)

Arguments

name
the name of the report

id
the id of the report, if omitted, use the id of the latest report

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

direction
A string indicating if we want to move up or down the tree permitted values are upstream, downstream

propagate
A boolean indicating if we want to propogate out of date through the tree

max_depth
A numeric, how far back should the tree go, this can be useful to truncate a very large tree. (default = 100)

show_all
A boolean, should we show all reports in the tree, not just the latest.

Value

An orderly Tree object with the root corresponding to the given report.

Details

1. We have updated a report and now need to re-run everything that depends on it. 2. We have a report that we want to re-run to ensure uses the latest information. Which other reports are used (directly or indirectly) by this report?

This function display this information in an easily readable format. Allowing users to see the dependency tree and which reports are out of date and need to be re-run.

Remark

The tree is built using data from the local report database (see orderly_commit). This means that it will not find changes from a report that has not be run and commited. _i.e._ if a user changes a report to use or create different artefacts this will not be picked up by the function until the reports are re-run and commited to the archive.

Examples

path <- orderly::orderly_example("demo")

id <- orderly::orderly_run("other", root = path, parameters=list(nmin=0))

#> [ name       ]  other

#> [ id         ]  20200302-113143-0d640b04

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0

#> [ data       ]  source => extract: 20 x 2

#> [ start      ]  2020-03-02 11:31:43

#> 
#> > extract$number <- extract$number + rnorm(1)
#> 
#> > write.csv(extract, "summary.csv", row.names = TRUE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > do_plot(extract)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2020-03-02 11:31:43

#> [ elapsed    ]  Ran report in 0.01134992 secs

#> [ artefact   ]  summary.csv: 01c70f6e51fab381154775c7e70ed7b3
#> [ ...        ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3

orderly::orderly_commit(id, root = path)

#> [ commit     ]  other/20200302-113143-0d640b04

#> [ copy       ]

#> [ import     ]  other:20200302-113143-0d640b04

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3d6e7da759/archive/other/20200302-113143-0d640b04"

id <- orderly::orderly_run("use_dependency", root = path)

#> [ name       ]  use_dependency

#> [ id         ]  20200302-113143-46c58384

#> [ depends    ]  other@20200302-113143-0d640b04:summary.csv -> incoming.csv

#> [ start      ]  2020-03-02 11:31:43

#> 
#> > d <- read.csv("incoming.csv", stringsAsFactors = FALSE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(d$number, d$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 
#> 
#> > info <- orderly::orderly_run_info()
#> 
#> > saveRDS(info, "info.rds")

#> [ end        ]  2020-03-02 11:31:43

#> [ elapsed    ]  Ran report in 0.01294231 secs

#> [ artefact   ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3
#> [ ...        ]  info.rds: bf6ad377efef4abca966a871582fa7e3

orderly::orderly_commit(id, root = path)

#> [ commit     ]  use_dependency/20200302-113143-46c58384

#> [ copy       ]

#> [ import     ]  use_dependency:20200302-113143-46c58384

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3d6e7da759/archive/use_dependency/20200302-113143-46c58384"

id <- orderly::orderly_run("use_dependency_2", root = path)

#> [ name       ]  use_dependency_2

#> [ id         ]  20200302-113143-70fc285c

#> [ depends    ]  use_dependency@20200302-113143-46c58384:graph.png -> graph.png

#> [ start      ]  2020-03-02 11:31:43

#> 
#> > stopifnot(file.exists("graph.png"))
#> 
#> > saveRDS(orderly::orderly_run_info(), "third.rds")

#> [ end        ]  2020-03-02 11:31:43

#> [ elapsed    ]  Ran report in 0.001130104 secs

#> [ artefact   ]  third.rds: c2a7b5bf8fcf482833db1331e82ec32c

orderly::orderly_commit(id, root = path)

#> [ commit     ]  use_dependency_2/20200302-113143-70fc285c

#> [ copy       ]

#> [ import     ]  use_dependency_2:20200302-113143-70fc285c

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3d6e7da759/archive/use_dependency_2/20200302-113143-70fc285c"

orderly::orderly_graph("other", root = path)

#> other [20200302-113143-0d640b04]
#> +--use_dependency [20200302-113143-46c58384]
#>     +--use_dependency_2 [20200302-113143-70fc285c]
#> 

orderly::orderly_graph("use_dependency_2", root = path,
                                 direction = "upstream")

#> use_dependency_2 [20200302-113143-70fc285c]
#> +--use_dependency [20200302-113143-46c58384]
#>     +--other [20200302-113143-0d640b04]
#>