Download dependent reports

Download dependent reports from an orderly remote. This can only be used if the orderly_config.yml lists a remote. This allows for a centralised workflow where a central orderly store exists and holds the canonical copies of reports, from which versions can be downloaded into local stores.

orderly_pull_dependencies(name, root = NULL, locate = TRUE, remote = NULL)

orderly_pull_archive(
  name,
  id = "latest",
  root = NULL,
  locate = TRUE,
  remote = NULL
)

Arguments

name
Name of the report to download dependencies for

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

remote
Description of the location. Typically this is a character string indicating a remote specified in the remotes block of your orderly_config.yml. It is also possible to pass in a directly created remote object (e.g., using orderly_remote_path, or one provided by another package). If left NULL, then the default remote for this orderly repository is used - by default that is the first listed remote.

id
The identifier (for orderly_pull_archive). The default is to use the latest report.

Value

No return value, these functions are called only for their side effects

Details

The orderly_pull_archive function pulls report directly (without it being a dependent report).

After setting your username up you can run orderly_pull_dependencies("reportname") to pull the dependencies of "reportname" down so that "reportname" can be run, or you can run orderly_pull_archive("reportname") to pull a copy of "reportname" that has been run on the remote server.

Pulling an archive report from a remote also pulls its dependencies (recursively), and adds all of these to the local database. This may require migrating old orderly archives (orderly_migrate). Note that this migration will likely fail for remote orderly versions older than 0.6.8 because the migration needs to read data files on disk that are not included in the downloaded archive in order to collect all the information required for the database. In this case, ask the administrator of the remote orderly archive to migrate their archive, and then re-pull.

See also

Examples

# Suppose we have a "remote" orderly repository at some path.
# This might be read-only for you in practice and available via a
# network filesystem or a dropbox folder synced to your computer.
# We'll populate this with a pair of reports:
path_remote <- orderly::orderly_example("demo")
id <- orderly::orderly_run("other", list(nmin = 0),
                           root = path_remote, echo = FALSE)

#> [ name       ]  other

#> [ id         ]  20200302-113146-2d14a608

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0

#> [ data       ]  source => extract: 20 x 2

#> [ start      ]  2020-03-02 11:31:46

#> [ end        ]  2020-03-02 11:31:46

#> [ elapsed    ]  Ran report in 0.01105976 secs

#> [ artefact   ]  summary.csv: 01c70f6e51fab381154775c7e70ed7b3
#> [ ...        ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  other/20200302-113146-2d14a608

#> [ copy       ]

#> [ import     ]  other:20200302-113146-2d14a608

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3d6f448683/archive/other/20200302-113146-2d14a608"

id <- orderly::orderly_run("use_dependency",
                           root = path_remote, echo = FALSE)

#> [ name       ]  use_dependency

#> [ id         ]  20200302-113146-6833c370

#> [ depends    ]  other@20200302-113146-2d14a608:summary.csv -> incoming.csv

#> [ start      ]  2020-03-02 11:31:46

#> [ end        ]  2020-03-02 11:31:46

#> [ elapsed    ]  Ran report in 0.01138663 secs

#> [ artefact   ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3
#> [ ...        ]  info.rds: 8fe031d315c6de2c224d61f7e058c49a

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  use_dependency/20200302-113146-6833c370

#> [ copy       ]

#> [ import     ]  use_dependency:20200302-113146-6833c370

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3d6f448683/archive/use_dependency/20200302-113146-6833c370"

# We'll create a an object to interact with this remote using
# orderly_remote_path.
remote <- orderly::orderly_remote_path(path_remote)

# We can use this object directly
remote$list_reports()

#>  [1] "changelog"          "connection"         "global"            
#>  [4] "html"               "interactive"        "minimal"           
#>  [7] "multi-artefact"     "multifile-artefact" "other"             
#> [10] "slow1"              "slow10"             "slow3"             
#> [13] "spaces"             "use_dependency"     "use_dependency_2"  
#> [16] "use_resource"       "use_resource_dir"   "view"              

remote$list_versions("other")

#> [1] "20200302-113146-2d14a608"

# More typically one will interact with the functions
# orderly_pull_archive and orderly_pull_dependencies.

# Now, suppose that you have your "local" copy of this; it shares
# the same source (ordinarily these would both be under version
# control with git):
path_local <- orderly::orderly_example("demo")

# If we wanted to run the report "use_dependency" we need to have
# a copy of the report "other", on which it depends:
try(orderly::orderly_run("use_dependency", root = path_local))

#> [ name       ]  use_dependency

#> Error in orderly_find_report(id, name, config, draft = use_draft, must_work = TRUE) : 
#>   Did not find archive report other:latest

# We can "pull" depenencies of a report before running
orderly::orderly_pull_dependencies("use_dependency", remote = remote,
                                   root = path_local)

#> [ depends    ]  use_dependency has 1 dependency

#> [ pull       ]  other:20200302-113146-2d14a608

#> [ import     ]  other:20200302-113146-2d14a608

# Now we can run the report because we have a local copy of the
# dependency:
orderly::orderly_run("use_dependency", root = path_local)

#> [ name       ]  use_dependency

#> [ id         ]  20200302-113146-c0de037a

#> [ depends    ]  other@20200302-113146-2d14a608:summary.csv -> incoming.csv

#> [ start      ]  2020-03-02 11:31:46

#> 
#> > d <- read.csv("incoming.csv", stringsAsFactors = FALSE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(d$number, d$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 
#> 
#> > info <- orderly::orderly_run_info()
#> 
#> > saveRDS(info, "info.rds")

#> [ end        ]  2020-03-02 11:31:46

#> [ elapsed    ]  Ran report in 0.01201868 secs

#> [ artefact   ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3
#> [ ...        ]  info.rds: 6aa46dab383907b76cc4f8f36451fc5d

#> [1] "20200302-113146-c0de037a"

# We can also directly pull previously run reports:
orderly::orderly_pull_archive("use_dependency", id, remote = remote,
                              root = path_local)

#> [ pull       ]  use_dependency:20200302-113146-6833c370

#> [ depends    ]  other/20200302-113146-2d14a608

#> [ pull       ]  other:20200302-113146-2d14a608 already exists, skipping

#> [ import     ]  use_dependency:20200302-113146-6833c370

orderly::orderly_list_archive(root = path_local)

#>             name                       id
#> 1          other 20200302-113146-2d14a608
#> 2 use_dependency 20200302-113146-6833c370