Download dependent reports

Download dependent reports.

orderly_pull_dependencies(name, root = NULL, locate = TRUE,
  remote = NULL)

orderly_pull_archive(name, id = "latest", root = NULL, locate = TRUE,
  remote = NULL)

Arguments

name
Name of the report to download dependencies for

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE).

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

remote
Description of the location. Typically this is a character string indicating a remote specified in the remotes block of your orderly_config.yml. It is also possible to pass in a directly created remote object (e.g., using orderly_remote_path, or one provided by another package). If left NULL, then the default remote for this orderly repository is used - by default that is the first listed remote.

id
The identifier (for orderly_pull_archive. The default is to use the latest report.

Details

The orderly_pull_archive function pulls report directly (without it being a dependent report).

After setting your username up you can run orderly_pull_dependencies("reportname") to pull the dependencies of "reportname" down so that "reportname" can be run, or you can run orderly_pull_archive("reportname") to pull a copy of "reportname" that has been run on the remote server.

Pulling an archive report from a remote also pulls its dependencies (recursively), and adds all of these to the local database. This may require migrating old orderly archives (orderly_migrate). Note that this migration will likely fail for remote orderly versions older than 0.6.8 because the migration needs to read data files on disk that are not included in the downloaded archive in order to collect all the information required for the database. In this case, ask the administrator of the remote orderly archive to migrate their archive, and then re-pull.

See also

Examples

# Suppose we have a "remote" orderly repository at some path.
# This might be read-only for you in practice and available via a
# network filesystem or a dropbox folder synced to your computer.
# We'll populate this with a pair of reports:
path_remote <- orderly::orderly_example("demo")
id <- orderly::orderly_run("other", list(nmin = 0),
                           root = path_remote, echo = FALSE)

#> [ name       ]  other

#> [ id         ]  20190813-162455-c190cd38

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0

#> [ data       ]  source => extract: 20 x 2

#> [ start      ]  2019-08-13 16:24:55

#> [ end        ]  2019-08-13 16:24:55

#> [ elapsed    ]  Ran report in 0.01265097 secs

#> [ artefact   ]  summary.csv: 23877598325ab7d45999349ef868a7f5
#> [ ...        ]  graph.png: 7360cb2eed3327ff8a677b3598ed7343

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  other/20190813-162455-c190cd38

#> [ copy       ]

#> [ import     ]  other:20190813-162455-c190cd38

#> [ success    ]  :)

#> [1] "/tmp/RtmpC4lmv4/filed5bf036c4/archive/other/20190813-162455-c190cd38"

id <- orderly::orderly_run("use_dependency",
                           root = path_remote, echo = FALSE)

#> [ name       ]  use_dependency

#> [ id         ]  20190813-162456-0857182b

#> [ depends    ]  other@20190813-162455-c190cd38:summary.csv -> incoming.csv

#> [ start      ]  2019-08-13 16:24:56

#> [ end        ]  2019-08-13 16:24:56

#> [ elapsed    ]  Ran report in 0.01247144 secs

#> [ artefact   ]  graph.png: 7360cb2eed3327ff8a677b3598ed7343

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  use_dependency/20190813-162456-0857182b

#> [ copy       ]

#> [ import     ]  use_dependency:20190813-162456-0857182b

#> [ success    ]  :)

#> [1] "/tmp/RtmpC4lmv4/filed5bf036c4/archive/use_dependency/20190813-162456-0857182b"

# We'll create a an object to interact with this remote using
# orderly_remote_path.
remote <- orderly::orderly_remote_path(path_remote)

# We can use this object directly
remote$list_reports()

#>  [1] "changelog"          "connection"         "global"            
#>  [4] "html"               "interactive"        "minimal"           
#>  [7] "multi-artefact"     "multifile-artefact" "other"             
#> [10] "spaces"             "use_dependency"     "use_resource"      
#> [13] "use_resource_dir"   "view"              

remote$list_versions("other")

#> [1] "20190813-162455-c190cd38"

# More typically one will interact with the functions
# orderly_pull_archive and orderly_pull_dependencies.

# Now, suppose that you have your "local" copy of this; it shares
# the same source (ordinarily these would both be under version
# control with git):
path_local <- orderly::orderly_example("demo")

# If we wanted to run the report "use_dependency" we need to have
# a copy of the report "other", on which it depends:
try(orderly::orderly_run("use_dependency", root = path_local))

#> [ name       ]  use_dependency

#> Error in orderly_latest(name, config, FALSE, draft = draft, must_work = must_work) : 
#>   Did not find any archive reports for other

# We can "pull" depenencies of a report before running
orderly::orderly_pull_dependencies("use_dependency", remote = remote,
                                   root = path_local)

#> [ depends    ]  use_dependency has 1 dependency

#> [ pull       ]  other:20190813-162455-c190cd38

#> [ import     ]  other:20190813-162455-c190cd38

# Now we can run the report because we have a local copy of the
# dependency:
orderly::orderly_run("use_dependency", root = path_local)

#> [ name       ]  use_dependency

#> [ id         ]  20190813-162456-83eb4c2a

#> [ depends    ]  other@20190813-162455-c190cd38:summary.csv -> incoming.csv

#> [ start      ]  2019-08-13 16:24:56

#> 
#> > d <- read.csv("incoming.csv", stringsAsFactors = FALSE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(d$number, d$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2019-08-13 16:24:56

#> [ elapsed    ]  Ran report in 0.01363182 secs

#> [ artefact   ]  graph.png: 7360cb2eed3327ff8a677b3598ed7343

#> [1] "20190813-162456-83eb4c2a"

# We can also directly pull previously run reports:
orderly::orderly_pull_archive("use_dependency", id, remote = remote,
                              root = path_local)

#> [ pull       ]  use_dependency:20190813-162456-0857182b

#> [ depends    ]  other/20190813-162455-c190cd38

#> [ pull       ]  other:20190813-162455-c190cd38 already exists, skipping

#> [ import     ]  use_dependency:20190813-162456-0857182b

orderly::orderly_list_archive(root = path_local)

#>             name                       id
#> 1          other 20190813-162455-c190cd38
#> 2 use_dependency 20190813-162456-0857182b