Orderly remote at a different path

Create a "handle" for interacting with orderly repositories that are hosted at a different path. This might be useful in cases where you have access to an orderly repository via a network mount or a synchronised folder (e.g., Dropbox, Box, etc). More generally, orderly_remote_path implements an interface used by orderly to abstract over different ways that orderly repositories might be hosted remotely, including over HTTP APIs.

orderly_remote_path(path, name = NULL)

Arguments

path
Path to the orderly store

name
Name of the remote

Value

An orderly_remote_path object, with methods that orderly will use in order to control this remote

See also

Examples

# Suppose we have a "remote" orderly repository at some path.
# This might be read-only for you in practice and available via a
# network filesystem or a dropbox folder synced to your computer.
# We'll populate this with a pair of reports:
path_remote <- orderly::orderly_example("demo")
id <- orderly::orderly_run("other", list(nmin = 0),
                           root = path_remote, echo = FALSE)

#> [ name       ]  other

#> [ id         ]  20200302-113147-c7286166

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0

#> [ data       ]  source => extract: 20 x 2

#> [ start      ]  2020-03-02 11:31:47

#> [ end        ]  2020-03-02 11:31:47

#> [ elapsed    ]  Ran report in 0.01120353 secs

#> [ artefact   ]  summary.csv: 01c70f6e51fab381154775c7e70ed7b3
#> [ ...        ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  other/20200302-113147-c7286166

#> [ copy       ]

#> [ import     ]  other:20200302-113147-c7286166

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3dbce58dc/archive/other/20200302-113147-c7286166"

id <- orderly::orderly_run("use_dependency",
                           root = path_remote, echo = FALSE)

#> [ name       ]  use_dependency

#> [ id         ]  20200302-113148-02006812

#> [ depends    ]  other@20200302-113147-c7286166:summary.csv -> incoming.csv

#> [ start      ]  2020-03-02 11:31:48

#> [ end        ]  2020-03-02 11:31:48

#> [ elapsed    ]  Ran report in 0.01158094 secs

#> [ artefact   ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3
#> [ ...        ]  info.rds: bbed2688e66f29e35beffa82d851122f

orderly::orderly_commit(id, root = path_remote)

#> [ commit     ]  use_dependency/20200302-113148-02006812

#> [ copy       ]

#> [ import     ]  use_dependency:20200302-113148-02006812

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3dbce58dc/archive/use_dependency/20200302-113148-02006812"

# We'll create a an object to interact with this remote using
# orderly_remote_path.
remote <- orderly::orderly_remote_path(path_remote)

# We can use this object directly
remote$list_reports()

#>  [1] "changelog"          "connection"         "global"            
#>  [4] "html"               "interactive"        "minimal"           
#>  [7] "multi-artefact"     "multifile-artefact" "other"             
#> [10] "slow1"              "slow10"             "slow3"             
#> [13] "spaces"             "use_dependency"     "use_dependency_2"  
#> [16] "use_resource"       "use_resource_dir"   "view"              

remote$list_versions("other")

#> [1] "20200302-113147-c7286166"

# More typically one will interact with the functions
# orderly_pull_archive and orderly_pull_dependencies.

# Now, suppose that you have your "local" copy of this; it shares
# the same source (ordinarily these would both be under version
# control with git):
path_local <- orderly::orderly_example("demo")

# If we wanted to run the report "use_dependency" we need to have
# a copy of the report "other", on which it depends:
try(orderly::orderly_run("use_dependency", root = path_local))

#> [ name       ]  use_dependency

#> Error in orderly_find_report(id, name, config, draft = use_draft, must_work = TRUE) : 
#>   Did not find archive report other:latest

# We can "pull" depenencies of a report before running
orderly::orderly_pull_dependencies("use_dependency", remote = remote,
                                   root = path_local)

#> [ depends    ]  use_dependency has 1 dependency

#> [ pull       ]  other:20200302-113147-c7286166

#> [ import     ]  other:20200302-113147-c7286166

# Now we can run the report because we have a local copy of the
# dependency:
orderly::orderly_run("use_dependency", root = path_local)

#> [ name       ]  use_dependency

#> [ id         ]  20200302-113148-5d1e43c8

#> [ depends    ]  other@20200302-113147-c7286166:summary.csv -> incoming.csv

#> [ start      ]  2020-03-02 11:31:48

#> 
#> > d <- read.csv("incoming.csv", stringsAsFactors = FALSE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(d$number, d$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 
#> 
#> > info <- orderly::orderly_run_info()
#> 
#> > saveRDS(info, "info.rds")

#> [ end        ]  2020-03-02 11:31:48

#> [ elapsed    ]  Ran report in 0.01215601 secs

#> [ artefact   ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3
#> [ ...        ]  info.rds: 11d51fca6570c9f279124229eaf56efd

#> [1] "20200302-113148-5d1e43c8"

# We can also directly pull previously run reports:
orderly::orderly_pull_archive("use_dependency", id, remote = remote,
                              root = path_local)

#> [ pull       ]  use_dependency:20200302-113148-02006812

#> [ depends    ]  other/20200302-113147-c7286166

#> [ pull       ]  other:20200302-113147-c7286166 already exists, skipping

#> [ import     ]  use_dependency:20200302-113148-02006812

orderly::orderly_list_archive(root = path_local)

#>             name                       id
#> 1          other 20200302-113147-c7286166
#> 2 use_dependency 20200302-113148-02006812