Run a report

Run a report. This will create a new directory in drafts/<reportname>, copy your declared resources there, extract data from databases (if you are using them), run your script and check that all expected artefacts were created. Once successfully run you can use orderly_commit to move it to the archive directory.

orderly_run(
  name,
  parameters = NULL,
  envir = NULL,
  root = NULL,
  locate = TRUE,
  echo = TRUE,
  id_file = NULL,
  fetch = FALSE,
  ref = NULL,
  message = NULL,
  instance = NULL,
  use_draft = FALSE,
  remote = NULL
)

Arguments

name
Name of the report to run (see orderly_list). A leading src/ will be removed if provided, allowing easier use of autocomplete.

parameters
Parameters passed to the report. A named list of parameters declared in the orderly.yml. Each parameter must be a scalar character, numeric, integer or logical.

envir
The parent of the environment that will be used to evaluate the report script; by default a new environment will be made with the global environment as the parent.

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

echo
Print the result of running the R code to the console

id_file
Write the identifier into a file

fetch
Logical, indicating if git should be fetched before checking out the reference ref.

ref
A git reference to use for this run (see Details)

message
An optional character string containing a message explaining why the report was run

instance
Select instance of the source database to be used, where multiple instances are configured. Use a single unnamed character string to indicate an instance to match. If given, then this name must be present in all databases where instances are listed in orderly_config.yml, and will be ignored by all database where instances are not given. See the "orderly" vignette for further information.

use_draft
Should draft reports be used for dependencies? This should be used only in development. Valid values are logical (TRUE, FALSE) or use the string newer to use draft reports where they are newer than archive reports. For consistency, always and never are equivalent to TRUE and FALSE, respectively.

remote
Remote to use to resolve dependencies. Use this in order to run a report with the same dependencies as are available on a remote server, particularly when using id = "latest". Note that this is not the same as running orderly_pull_dependencies, then orderly_run with remote = NULL, as the pull/run approach will use the latest report in your archive but the remote = "remote" approach will use the latest approach in the remote archive (which might be less recent).

Value

The id of the newly created report

Details

If ref is provided then before running a report orderly will try to check out (as a detached HEAD) ref, interpreted as a git reference. This can be a commit, tag, or a branch name (including remote). The working directory must be clean according to git status and this will require some careful use of .gitignore to exclude draft, archive, data and orderly.sqlite. The git tree will revert back to the original branch at completion (or failure to complete) the report.

Parameters are passed to the report as a named list, for example

id <- orderly::orderly_run("other", list(nmin = 0.2), root = path)

(see the examples). The names of the parameters (here, nmin) must correspond to declared parameters in the orderly.yml. It is an error if parameters without a default are omitted, and it is an error if unknown parameters are provided.

See also

Examples

path <- orderly::orderly_example("demo")

# To run most reports, provide the report name (and the path if
# not running in the working directory, as is the case here):
id <- orderly::orderly_run("minimal", root = path)

#> [ name       ]  minimal

#> [ id         ]  20200302-113148-b570178d

#> [ data       ]  source => dat: 20 x 2

#> [ start      ]  2020-03-02 11:31:48

#> 
#> > png("mygraph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(dat$number, dat$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2020-03-02 11:31:48

#> [ elapsed    ]  Ran report in 0.01063013 secs

#> [ artefact   ]  mygraph.png: 5f8e1c71677703144642256f70c86502

# Every report gets a unique identifier, based on the time (it is
# ISO 8601 time with random hex appended to end)
id

#> [1] "20200302-113148-b570178d"

# After being run, a report is a "draft" and will exist in the
# drafts directory:
orderly::orderly_list_drafts(root = path)

#>      name                       id
#> 1 minimal 20200302-113148-b570178d

# Draft reports are always stored in the path
# <root>/draft/<name>/<id>, so we have
dir(file.path(path, "draft", "minimal", id))

#> [1] "mygraph.png"     "orderly_run.rds" "orderly.yml"     "script.R"       

# which contains the files when the report was run.

# If a report has parameters, then these must be passed in as a
# named list.
id <- orderly::orderly_run("other", list(nmin = 0.2), root = path)

#> [ name       ]  other

#> [ id         ]  20200302-113148-c4eef58b

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0.2

#> [ data       ]  source => extract: 19 x 2

#> [ start      ]  2020-03-02 11:31:48

#> 
#> > extract$number <- extract$number + rnorm(1)
#> 
#> > write.csv(extract, "summary.csv", row.names = TRUE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > do_plot(extract)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2020-03-02 11:31:48

#> [ elapsed    ]  Ran report in 0.01295972 secs

#> [ artefact   ]  summary.csv: da1bdb89343fc0c1c9b19f3d257d27f7
#> [ ...        ]  graph.png: 0f5857002852d4d85f0d3423907eeb2c

# These parameters can be used in SQL queries or in the report
# code.