Run a report

Run a report. The orderly_data function is for testing the queries (and developing the report).

orderly_run(name, parameters = NULL, envir = NULL, root = NULL,
  locate = TRUE, echo = TRUE, id_file = NULL, fetch = FALSE,
  ref = NULL, message = NULL)

orderly_data(name, parameters = NULL, envir = NULL, root = NULL,
  locate = TRUE)

Arguments

name
Name of the report to run (see orderly_list).

parameters
Parameters passed to the report. A named list of parameters declared in the orderly.yml.

envir
The parent of environment to evaluate the report in; by default a new environment will be made with the global environment as the parent.

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE).

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

echo
Print the result of running the R code to the console

id_file
Write the identifier into a file

fetch
Logical, indicating if git should be fetched before checking out the reference ref.

ref
A git reference to use for this run (see Details)

message
An optional character string containing a message explaining why the report was run

Details

If ref is provided then before running a report orderly will try to check out (as a detached HEAD) ref, interpreted as a git reference. This can be a commit, tag, or a branch name (including remote). The working directory must be clean according to git status and this will require some careful use of .gitignore to exclude draft, archive, data and orderly.sqlite. The git tree will revert back to the original branch at completion (or failure to complete) the report.

See also

Examples

path <- orderly::orderly_example("demo")

# To run most reports, provide the report name (and the path if
# not running in the working directory, as is the case here):
id <- orderly::orderly_run("minimal", root = path)

#> [ name       ]  minimal

#> [ id         ]  20190813-162458-e414f551

#> [ data       ]  source => dat: 20 x 2

#> [ start      ]  2019-08-13 16:24:58

#> 
#> > png("mygraph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(dat$number, dat$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2019-08-13 16:24:58

#> [ elapsed    ]  Ran report in 0.01220465 secs

#> [ artefact   ]  mygraph.png: 7360cb2eed3327ff8a677b3598ed7343

# Every report gets a unique identifier, based on the time (it is
# ISO 8601 time with random hex appended to end)
id

#> [1] "20190813-162458-e414f551"

# After being run, a report is a "draft" and will exist in the
# drafts directory:
orderly::orderly_list_drafts(root = path)

#>      name                       id
#> 1 minimal 20190813-162458-e414f551

# Draft reports are always stored in the path
# <root>/draft/<name>/<id>, so we have
dir(file.path(path, "draft", "minimal", id))

#> [1] "mygraph.png"     "orderly_run.rds" "orderly.yml"     "script.R"       

# which contains the files when the report was run.

# If a report has parameters, then these must be passed in as a
# named list.
id <- orderly::orderly_run("other", list(nmin = 0.2), root = path)

#> [ name       ]  other

#> [ id         ]  20190813-162458-f9cc4de0

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0.2

#> [ data       ]  source => extract: 17 x 2

#> [ start      ]  2019-08-13 16:24:58

#> 
#> > write.csv(extract, "summary.csv", row.names = TRUE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > do_plot(extract)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2019-08-13 16:24:59

#> [ elapsed    ]  Ran report in 0.01221991 secs

#> [ artefact   ]  summary.csv: 4e4987a9bfc3a18ed9c3bfc54114610a
#> [ ...        ]  graph.png: dfea86de3fd70478dee822869fb8bd49

# These parameters can be used in SQL queries or in the report
# code.
# The function orderly_data does all the preparation work that
# orderly_run does, but does not run the report; instead it
# returns the created environment with all the data and parameters
# set.
env <- orderly::orderly_data("other", list(nmin = 0.2), root = path)

#> [ parameter  ]  nmin: 0.2

#> [ data       ]  source => extract: 17 x 2

ls(env)

#> [1] "extract" "nmin"   

env$nmin

#> [1] 0.2

env$extract

#>                                name    number
#> 1          displeasing_waterdragons 0.8209463
#> 2            frangible_caimanlizard 0.6470602
#> 3               nitpicking_nabarlek 0.7829328
#> 4                undivined_axisdeer 0.5530363
#> 5                     crazy_cricket 0.5297196
#> 6                   uncongested_gar 0.7893562
#> 7       potential_galapagostortoise 0.4772301
#> 8                    pasty_tapaculo 0.7323137
#> 9  auspicial_europeanfiresalamander 0.6927316
#> 10                    crooked_husky 0.4776196
#> 11              conservable_mallard 0.8612095
#> 12   psychopharmacological_ibisbill 0.4380971
#> 13                 gastric_bobolink 0.2447973
#> 14            releasable_sanderling 0.3162717
#> 15    worldlywise_asianwaterbuffalo 0.5186343
#> 16                  fungoid_pitbull 0.6620051
#> 17                  sightly_gonolek 0.4068302