Information on current orderly run

This function allows inspection of some of orderly's metadata during an orderly run. The format returned is internal to orderly and subject to change. It is designed to be used within report code. To use in conjunction with orderly_test_start, you must pass in the path to the report in question.

orderly_run_info(path = NULL)

Arguments

path
Path to the report currently being run. This should be left as NULL when running a report, and the path to the report being run should be used when using orderly_test_start

Value

A list of metadata about the current report

Examples

path <- orderly::orderly_example("demo")

# This example uses orderly_run_info within its script, saving the
# output to "output.rds"
readLines(file.path(path, "src", "use_dependency", "script.R"))

#> [1] "d <- read.csv(\"incoming.csv\", stringsAsFactors = FALSE)"
#> [2] ""                                                         
#> [3] "png(\"graph.png\")"                                       
#> [4] "par(mar = c(15, 4, .5, .5))"                              
#> [5] "barplot(setNames(d$number, d$name), las = 2)"             
#> [6] "dev.off()"                                                
#> [7] ""                                                         
#> [8] "info <- orderly::orderly_run_info()"                      
#> [9] "saveRDS(info, \"info.rds\")"                              

# Run the dependency:
id <- orderly::orderly_run("other", list(nmin = 0), root = path)

#> [ name       ]  other

#> [ id         ]  20200302-113148-fb9aae8e

#> [ sources    ]  functions.R

#> [ parameter  ]  nmin: 0

#> [ data       ]  source => extract: 20 x 2

#> [ start      ]  2020-03-02 11:31:48

#> 
#> > extract$number <- extract$number + rnorm(1)
#> 
#> > write.csv(extract, "summary.csv", row.names = TRUE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > do_plot(extract)

#> 
#> > dev.off()
#> pdf 
#>   2 

#> [ end        ]  2020-03-02 11:31:49

#> [ elapsed    ]  Ran report in 0.01117611 secs

#> [ artefact   ]  summary.csv: 01c70f6e51fab381154775c7e70ed7b3
#> [ ...        ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3

orderly::orderly_commit(id, root = path)

#> [ commit     ]  other/20200302-113148-fb9aae8e

#> [ copy       ]

#> [ import     ]  other:20200302-113148-fb9aae8e

#> [ success    ]  :)

#> [1] "/tmp/Rtmp3R7mks/file3d54d11aa6/archive/other/20200302-113148-fb9aae8e"

# Then the report
id <- orderly::orderly_run("use_dependency", root = path)

#> [ name       ]  use_dependency

#> [ id         ]  20200302-113149-39d5d84d

#> [ depends    ]  other@20200302-113148-fb9aae8e:summary.csv -> incoming.csv

#> [ start      ]  2020-03-02 11:31:49

#> 
#> > d <- read.csv("incoming.csv", stringsAsFactors = FALSE)
#> 
#> > png("graph.png")
#> 
#> > par(mar = c(15, 4, 0.5, 0.5))
#> 
#> > barplot(setNames(d$number, d$name), las = 2)

#> 
#> > dev.off()
#> pdf 
#>   2 
#> 
#> > info <- orderly::orderly_run_info()
#> 
#> > saveRDS(info, "info.rds")

#> [ end        ]  2020-03-02 11:31:49

#> [ elapsed    ]  Ran report in 0.01196003 secs

#> [ artefact   ]  graph.png: 9918b4caa5f89dff21a92e8a4fdeb7c3
#> [ ...        ]  info.rds: 1f4659146c08e14bc70eefd3e8113526

# This is the contents:
readRDS(file.path(path, "draft", "use_dependency", id, "info.rds"))

#> $data
#> NULL
#> 
#> $script
#> [1] "script.R"
#> 
#> $artefacts
#>      filenames   description                          format       
#> [1,] "graph.png" "A summary graph"                    "staticgraph"
#> [2,] "info.rds"  "Example of orderly_run_info output" "data"       
#> 
#> $depends
#>                         id  name draft    filename           as is_pinned index
#> 1 20200302-113148-fb9aae8e other    NA summary.csv incoming.csv     FALSE     1
#>                                                                    path
#> 1 /tmp/Rtmp3R7mks/file3d54d11aa6/archive/other/20200302-113148-fb9aae8e
#>   is_latest                             hash                time id_requested
#> 1      TRUE 01c70f6e51fab381154775c7e70ed7b3 2020-03-02 11:31:49       latest
#> 
#> $displayname
#> [1] "Report using a dependency"
#> 
#> $author
#> [1] "Dr Very Serious"
#> 
#> $requester
#> [1] "ACME"
#> 
#> $comment
#> [1] NA
#> 
#> $name
#> [1] "use_dependency"
#> 
#> $script_hash
#> [1] "fd8a8beb696d62013e70ecb77f670052"
#> 
#> $hash
#> [1] "e0b738e70a3b1d0e7404ddb2984cdd18"
#> 
#> $path
#> [1] "/tmp/Rtmp3R7mks/file3d54d11aa6/src/use_dependency"
#> 
#> $id
#> [1] "20200302-113149-39d5d84d"
#> 
#> $workdir
#> [1] "/tmp/Rtmp3R7mks/file3d54d11aa6/draft/use_dependency/20200302-113149-39d5d84d"
#> 
#> $owd
#> [1] "/orderly/docs/reference"
#> 
#> $inputs
#>   file_purpose    filename                        file_hash file_size
#> 1  orderly_yml orderly.yml 146a5fc70197dfdb3aaf54586dcec6ee       358
#> 2       script    script.R fd8a8beb696d62013e70ecb77f670052       220
#>