Prepare a directory for orderly to use

For interactive testing of orderly code. This runs through and sets everything up as orderly would (creates a new working directory and changes into it, pulls data from the database, copies over any dependent reports) but then rather than running the report hands back to the user. The prompt looks like browser but it is just a plain old R prompt and the code runs in the global environment.

orderly_test_start(name, parameters = NULL, envir = .GlobalEnv,
  root = NULL, locate = TRUE)

orderly_test_end(cleanup = FALSE)

orderly_test_restart(cleanup = TRUE)

orderly_test_check()

Arguments

name
Name of the report to run (see orderly_list).

parameters
Parameters passed to the report. A named list of parameters declared in the orderly.yml.

envir
The parent of environment to evaluate the report in; by default a new environment will be made with the global environment as the parent.

root
The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE).

locate
Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an orderly_config.yml file.

cleanup
Delete testing directory on exit? If FALSE then you will probably want to use orderly_cleanup later to delete the test directory. Note that it is not possible to commit the results of an orderly test run

Details

To quit run orderly_test_end() (or enter Q, like browser). To test if all artefacts have been created run orderly_test_check().

Examples

path <- orderly::orderly_example("minimal")
orderly::orderly_test_start("example", root = path)

#> [ name       ]  example

#> [ id         ]  20190813-162459-ec63be4e

#> [ data       ]  source => dat: 20 x 2

#> [ setwd      ]  running in test draft directory

# R is now running from the newly created draft directory for this
# report:
getwd()

#> [1] "/tmp/RtmpC4lmv4/filed10ca3b98/draft/example/20190813-162459-ec63be4e"

# The data in the orderly example is now available to use
dat

#>                                name     number
#> 1          displeasing_waterdragons 0.82094629
#> 2            frangible_caimanlizard 0.64706019
#> 3               nitpicking_nabarlek 0.78293276
#> 4                undivined_axisdeer 0.55303631
#> 5                     crazy_cricket 0.52971958
#> 6                   uncongested_gar 0.78935623
#> 7    unsuspecting_africanclawedfrog 0.02333120
#> 8       potential_galapagostortoise 0.47723007
#> 9                    pasty_tapaculo 0.73231374
#> 10 auspicial_europeanfiresalamander 0.69273156
#> 11                    crooked_husky 0.47761962
#> 12              conservable_mallard 0.86120948
#> 13   psychopharmacological_ibisbill 0.43809711
#> 14                 gastric_bobolink 0.24479728
#> 15                  serpentine_slug 0.07067905
#> 16             kimberlite_nighthawk 0.09946616
#> 17            releasable_sanderling 0.31627171
#> 18    worldlywise_asianwaterbuffalo 0.51863426
#> 19                  fungoid_pitbull 0.66200508
#> 20                  sightly_gonolek 0.40683019

# Check to see which artefacts have been created so far:
orderly::orderly_test_check()

#> [ artefact   ]  mygraph.png: <missing>

# Manually the code that this report has in its script
png("mygraph.png")
par(mar = c(15, 4, .5, .5))
barplot(setNames(dat$number, dat$name), las = 2)
dev.off()

#> pdf 
#>   2 

orderly::orderly_test_check()

#> [ artefact   ]  mygraph.png: 7360cb2eed3327ff8a677b3598ed7343

# Revert back to the original directory:
orderly::orderly_test_end()

#> [ setwd      ]  reverting to original directory